+++ app/models/event.rb
@@ -65,8 +65,8 @@ class Event < ApplicationRecord
   end
 
   def guild_role_of_user(user_id)
-    return nil if eventable_type != 'Guild' && eventable_type != 'Static'
-    return nil if eventable_type == 'Static' && eventable.staticable_type != 'Guild'
+    return nil if eventable_type == 'World'
+    return nil if eventable_type == 'Static' && eventable.staticable_type == 'Character'
     guild = eventable_type == 'Static' ? eventable.staticable : eventable
     # leaders from guild of this user
     leaders = guild.characters_with_leader_role.select { |character| character.user_id == user_id }