+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Fixed
 - bug with approving to static event of guild by Class Leader
+- bug with updating subscribe by Class Leader
 
 ## [0.8.3] - 2019-11-18
 ### Fixed